+++ ScanOrchestrator.cs
@@ -101,6 +101,7 @@ internal static class ScanOrchestrator
         if (!opts.NoUsn)       s.Add(("NTFS USN journal (deleted files)",  sync(sec => UsnJournalScanner.Run(sec))));
         if (!opts.NoDefender)  s.Add(("Windows Defender history",          sync(sec => DefenderLogScanner.Run(sec))));
         if (!opts.NoEngines)   s.Add(("Heuristic engines",                 sync(sec => HeuristicEngineScanner.Run(report, sec))));
+        if (!opts.NoAgentScan) s.Add(("Java-agent manifest scan",          sync(sec => JavaAgentScanner.Run(sec))));
         if (!opts.NoExeScan)   s.Add(("Renamed-cheat detector (hash + content markers)",
                                                                             sync(sec => CheatExeScanner.Run(sec))));